+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to the versioning scheme outlined in the [README.md](RE
 
 - The HTTP `Date` header in responses now strictly follows RFC7231.
 - When a previous block commit is unable to be RBFed, the miner will now just wait for it to be confirmed instead of submitting a new block commit which breaks the miner's UTXO chain.
+- Attempt heuristically to avoid re-sending stackerdb chunk events that have already been sent to observers
 
 ### Fixed